Finished the method that loads the opacity transfer functions from disk and creates their textures on the GPU memory. At least the loading and uploading of the multidimensional opacity transfer functions should be functioning now.

+++ Kernel/multidimvolumedata.h
@@ -14,7 +14,7 @@ namespace ggraf
         typedef struct
         {
             std::string path;
-            size_t w, h;
+            int w, h;
             size_t bytes_per_pixel;
         } ParsedMultiDimTFPath;
